Move QA class to its own library for code sharing
